requirements.txt
attrs==26.1.0
autobahn==19.11.2
Automat==25.4.16
beautifulsoup4==4.14.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
constantly==23.10.4
cryptography==46.0.7
curl_cffi==0.15.0
frozendict==2.4.7
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
kiteconnect==5.1.0
markdown-it-py==4.0.0
mdurl==0.1.2
multitasking==0.0.12
numpy==2.4.4
packaging==26.0
pandas==3.0.2
peewee==4.0.4
platformdirs==4.9.6
protobuf==7.34.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
Pygments==2.20.0
pyOpenSSL==26.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
requests==2.33.1
rich==15.0.0
service-identity==24.2.0
six==1.17.0
soupsieve==2.8.3
streamlit==1.56.0
Twisted==25.5.0
txaio==25.12.2
typing_extensions==4.15.0
urllib3==2.6.3
websockets==16.0
yfinance==1.2.2
zope.interface==8.3
